requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.3
Flask-WTF==1.2.1
Flask-Migrate==4.0.5
Flask-Admin==1.6.1
WTForms==3.1.1
email-validator==2.1.0
python-dotenv==1.0.0
Pillow==12.1.1
requests==2.32.5
bleach==6.3.0
